Replaced method chaining with static tasks and imrpoved documentation + did some minor changes.

+++ WindowsOptimizations.Core/Extensions/StringExtensions.cs
@@ -8,7 +8,10 @@ namespace WindowsOptimizations.Core.Extensions
     /// </summary>
     public enum SizeUnits
     {
-        Byte, KB, MB, GB,
+        Byte,
+        KB,
+        MB,
+        GB,
     }
 
     /// <summary>
@@ -19,12 +22,12 @@ namespace WindowsOptimizations.Core.Extensions
         /// <summary>
         /// Formats the value to a readable storage size such as MB.
         /// </summary>
-        /// <param name="value"></param>
-        /// <param name="unit"></param>
-        /// <returns></returns>
+        /// <param name="value">The unmodified value.</param>
+        /// <param name="unit">The preferred unit to convert the value.</param>
+        /// <returns>[<see cref="string"/>] The modified value according to the specified size unit.</returns>
         public static string ToSize(long value, SizeUnits unit)
         {
-            return (value / (double)Math.Pow(1024, (long)unit)).ToString("0.00");
+            return (value / Math.Pow(1024, (long)unit)).ToString("0.00");
         }
     }
 }
